Introduce kahpp-spring-autoconfigure module
This first module is the most important since it contains the entire kahpp core, including the test suite.
It's usabilla/kahpp, with additional configuration to make it usable by the spring boot starter.
The main difference is the test suite, which is now moved inside `main` folder to make it usable by other projects.