Fix handle by status code on HTTP error
Handlers on Kahpp are implemented in a way that doesn't consent us from preserving the HTTP response in case of error easily.
On Kahpp, a standard HTTP step in case of error puts the HTTP response in the record's header and then routes the message on the error topic.
All Handlers work fine with this, except RECORD_ROUTE.
The goal of RECORD_ROUTE is to route a record based on the HTTP status code. If the status code is an error, we should preserve the response in the header, but we lost the response because we treated the error like a successful HTTP request, which is wrong.

For now, this fix helps to don't miss the HTTP response, so we'll have the errors in the header also for RECORD_ROUTE, like the default for other HTTP steps.
Anyway, this functionality (Handlers) needs to be refactored.

Signed-off-by: Stefano Guerrini <sguerrini@surveymonkey.com>