Suppress warning where synchronized is not needed
From 'UnsynchronizedOverridesSynchronized' errorprone doc:
'''
Thread-safe methods should never be overridden by methods that are not thread-safe.
Doing so violates behavioural subtyping, and can result in bugs if the subtype is used in contexts that rely on the thread-safety of the supertype.
'''

Avoid using the 'synchronised' keyword on this overridden method is acceptable;
Values come from a static configuration and never change over time.

Signed-off-by: Stefano Guerrini <sguerrini@surveymonkey.com>