Introduce 'to' field on Split step
Before this addition, the 'Split' step was overriding the value field.
The main reason behind the 'to' field is to have a way to define where to put the extracted object.
To keep backward compatibility, the default value of 'to' is an empty string, so if 'to' is not part of the configuration, we can be sure that the step will work like before.

To make this minor update was necessary a little workaround (To don't break other functionalities), but it became clear that KaHPPRecord and all Mutation classes need a big refactor.

Signed-off-by: Stefano Guerrini <sguerrini@surveymonkey.com>