Introduce dynamic path for HTTP Call
Let's say that we want to call an endpoint using a different value on the path for each record, with this implementation we can easily do that.
Using a placeholder we can take data from the record a mutate the path dynamically.

Example

We have a list of producs:
```
[
    {
        "productId": 1
    },
    {
        "productId": 2
    },
    {
        "productId": 3
    },
    {
        "productId": 4
    }
]
```

We can receive the info about each product by calling '/products/<id of the product>'

To do that we use this config:
```
    config:
      api: pruduct-api
      path: /products/${value.productId}
      topic: error
```

If the value is not there, the call will fail, and the record will be routed to the error topic.

Signed-off-by: Stefano Guerrini <sguerrini@surveymonkey.com>