Fix path placeholder overriding itself
The new placeholder functionality, was introduced with a bug.
After the first replacement of the placeholder, the path was overriding.
The result is that all calls were the same.
With the introduction of `actualPath`, this issue should be resolved.
`path` will remain the same, keeping the placeholder, when 'actualPath' will change for each record.

Signed-off-by: Stefano Guerrini <sguerrini@surveymonkey.com>