Update dependency check suppression list
Outdated entries have been removed while new ones have been added so
that the GitHub dependency-check action passes again.
Note that the remaining suppressions include false positives,
vulnerabilities that have not been fixed yet or are a result of version
constraints of the (components of the) framework we use (Spring).

Signed-off-by: Clara van Miert <cvanmiert@momentive.ai>