Rename metric prefix
After the introduction of 'micrometer-prometheus' the exported metrics changed in the name.
So, for all Kahpp custom metrics, the prefix was 'kahppMetricPrefix' then the JMX exporter was mapping these to 'kahpp_*'.
Now that we use 'micrometer' we don't need to use JMX exporter, for this reason, the prefix will become 'kahpp'.

Signed-off-by: Stefano Guerrini <sguerrini@surveymonkey.com>